home_vision/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
psycopg[binary]==3.2.1
aiosqlite==0.20.0
orjson==3.10.7
pydantic-settings==2.6.1
python-dotenv==1.0.1
